requirements.txt
fastapi[all]
sqlalchemy
bcrypt
pydantic
motor
asyncpg
pyjwt